# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/products/{id}": {
      "get": {
        "operationId": "getAdminProductById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getAdminProductVariantGroups": {
                "operationId": "listAdminProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/id"
                }
              },
              "getPublicProductVariantGroups": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Products"
        ]
      },
      "patch": {
        "description": "Supports optional nested variantGroups[] updates: with id updates existing group/option, without id creates new one",
        "operationId": "updateAdminProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getUpdatedAdminProduct": {
                "operationId": "getAdminProductById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product",
        "tags": [
          "Admin Products"
        ]
      },
      "delete": {
        "operationId": "deleteAdminProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Admin Products"
        ]
      }
    },
...
    "/api/products/{productId}/variant-groups": {
      "get": {
        "operationId": "getPublicProductVariantGroups",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "calculatePriceFromFirstVariantOption": {
                "operationId": "getPublicProductVariantPrice",
                "parameters": {
                  "productId": "$request.path.productId",
                  "optionIds": "$response.body#/0/options/0/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductVariantGroupDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get active variant groups for product",
        "tags": [
          "Product Variants"
        ]
      }
    },
...
    "/api/branch-products": {
      "get": {
        "operationId": "getPublicBranchProducts",
        "parameters": [
          {
            "name": "branchId",
            "required": false,
            "in": "query",
            "description": "Branch id",
            "schema": {
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "price",
            "required": false,
            "in": "query",
            "description": "Price in branch",
            "schema": {
              "example": 25990,
              "type": "number"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "description": "Minimum price in branch",
            "schema": {
              "example": 10000,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "description": "Maximum price in branch",
            "schema": {
              "example": 50000,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Product name filter",
            "schema": {
              "example": "oak",
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": false,
            "in": "query",
            "description": "Product slug filter (partial match, case-insensitive)",
            "schema": {
              "example": "oak-dining",
              "type": "string"
            }
          },
          {
            "name": "description",
            "required": false,
            "in": "query",
            "description": "Product description filter",
            "schema": {
              "example": "dining",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "Product category id",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getBranchProductByIdFromList": {
                "operationId": "getPublicBranchProductById",
                "parameters": {
                  "id": "$response.body#/items/0/id"
                }
              },
              "getBranchProductBySlugFromList": {
                "operationId": "getPublicBranchProductBySlug",
                "parameters": {
                  "slug": "$response.body#/items/0/slug",
                  "branchId": "$response.body#/items/0/branchId"
                }
              },
              "getProductVariantGroupsFromBranchProductList": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/items/0/productId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchProductPaginationDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "BranchProduct"
        ]
      }
    },
    "/api/branch-products/by-slug": {
      "get": {
        "operationId": "getPublicBranchProductBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "query",
            "description": "Product slug",
            "schema": {
              "example": "oak-dining-table-120",
              "type": "string"
            }
          },
          {
            "name": "branchId",
            "required": true,
            "in": "query",
            "description": "Branch id",
            "schema": {
              "example": 5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getBranchProductByIdFromSlug": {
                "operationId": "getPublicBranchProductById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              },
              "getProductVariantGroupsFromBranchProductSlug": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/productId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchProductDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "BranchProduct"
        ]
      }
    },
    "/api/branch-products/{id}": {
      "get": {
        "operationId": "getPublicBranchProductById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getBranchProductBySlugFromId": {
                "operationId": "getPublicBranchProductBySlug",
                "parameters": {
                  "slug": "$response.body#/slug",
                  "branchId": "$response.body#/branchId"
                }
              },
              "getProductVariantGroupsFromBranchProductId": {
                "operationId": "getPublicProductVariantGroups",
                "parameters": {
                  "productId": "$response.body#/productId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchProductDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "BranchProduct"
        ]
      }
    },
...
    "/api/admin/branch-products/{id}": {
      "get": {
        "operationId": "getAdminBranchProductById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "links": {
              "getProductFromAdminBranchProduct": {
                "operationId": "getAdminProductById",
                "parameters": {
                  "id": "$response.body#/productId"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminBranchProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminBranchProducts"
        ]
      },
      "patch": {
        "operationId": "updateAdminBranchProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBranchProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "links": {
              "getUpdatedAdminBranchProduct": {
                "operationId": "getAdminBranchProductById",
                "parameters": {
                  "id": "$response.body#/id"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BranchProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminBranchProducts"
        ]
      },
      "delete": {
        "operationId": "deleteAdminBranchProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "AdminBranchProducts"
        ]
      }
    },
...
  },
  "info": {
    "title": "catalog-service API",
    "description": "REST API + GRPC endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002"
    }
  ],
  "components": {
    "schemas": {
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "Total items count",
            "example": 124
          },
          "page": {
            "type": "number",
            "description": "Current page",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Items per page",
            "example": 25
          }
        },
        "required": [
          "total",
          "page",
          "limit"
        ]
      },
...
      "ImageProductDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "description": "Image URL",
            "example": "https://cdn.example.com/products/oak-dining-table-120/main.jpg"
          },
          "type": {
            "type": "string",
            "description": "Image type",
            "example": "main"
          }
        },
        "required": [
          "url",
          "type"
        ]
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Product id",
            "example": 101
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Oak Dining Table 120"
          },
          "fullName": {
            "type": "string",
            "description": "Full product name",
            "example": "Solid Oak Dining Table 120 x 80 cm"
          },
          "sku": {
            "type": "string",
            "description": "SKU (article)",
            "example": "TBL-OAK-120"
          },
          "slug": {
            "type": "string",
            "description": "Product slug",
            "example": "oak-dining-table-120"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Natural solid oak dining table with matte finish"
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 24990
          },
          "categoryId": {
            "type": "number",
            "description": "Category id",
            "example": 3
          },
          "categoryName": {
            "type": "string",
            "description": "Category name",
            "example": "Dining Tables"
          },
          "sortOrder": {
            "type": "number",
            "description": "Sorting priority",
            "example": 10
          },
          "images": {
            "example": [
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/main.jpg",
                "type": "main"
              },
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/icon.jpg",
                "type": "icon"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageProductDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "price",
          "categoryId",
          "categoryName",
          "sortOrder",
          "images"
        ]
      },
...
      "UpdateProductVariantOptionInlineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant option id. If omitted, option will be created",
            "example": 201
          },
          "name": {
            "type": "string",
            "description": "Variant option title, for example 100 cm or oak",
            "example": "140 cm"
          },
          "priceDelta": {
            "type": "number",
            "description": "Price delta for this option",
            "example": 600
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 3
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "example": true
          }
        }
      },
      "UpdateProductVariantGroupInlineDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant group id. If omitted, group will be created",
            "example": 12
          },
          "name": {
            "type": "string",
            "description": "Variant group title, for example Size or Material",
            "example": "Size"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether one option from this group must be selected",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "example": true
          },
          "options": {
            "example": [
              {
                "id": 201,
                "priceDelta": 250,
                "sortOrder": 1
              },
              {
                "name": "160 cm",
                "priceDelta": 900,
                "sortOrder": 4,
                "isActive": true
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateProductVariantOptionInlineDto"
            }
          }
        }
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Oak Dining Table 140"
          },
          "fullName": {
            "type": "string",
            "description": "Full product name",
            "example": "Solid Oak Dining Table 140 x 80 cm"
          },
          "sku": {
            "type": "string",
            "description": "SKU (article)",
            "example": "TBL-OAK-140"
          },
          "slug": {
            "type": "string",
            "description": "Product slug",
            "example": "oak-dining-table-140"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Updated model with larger top and reinforced legs"
          },
          "price": {
            "type": "number",
            "description": "Product price",
            "example": 27990
          },
          "categoryId": {
            "type": "number",
            "description": "Category id",
            "example": 3
          },
          "sortOrder": {
            "type": "number",
            "description": "Sorting priority",
            "example": 12
          },
          "variantGroups": {
            "description": "Variant groups update payload. With id updates existing group, without id creates new one",
            "example": [
              {
                "id": 10,
                "name": "Size",
                "options": [
                  {
                    "id": 100,
                    "priceDelta": 250
                  },
                  {
                    "name": "140 cm",
                    "priceDelta": 600,
                    "sortOrder": 3
                  }
                ]
              },
              {
                "name": "Color",
                "isRequired": false,
                "options": [
                  {
                    "name": "White",
                    "priceDelta": 0
                  }
                ]
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpdateProductVariantGroupInlineDto"
            }
          }
        }
      },
...
      "ProductVariantOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant option id",
            "example": 201
          },
          "name": {
            "type": "string",
            "description": "Variant option title",
            "example": "120 cm"
          },
          "priceDelta": {
            "type": "number",
            "description": "Price delta added to base product price",
            "example": 300
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 2
          }
        },
        "required": [
          "id",
          "name",
          "priceDelta",
          "sortOrder"
        ]
      },
      "ProductVariantGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Variant group id",
            "example": 12
          },
          "name": {
            "type": "string",
            "description": "Variant group title, for example Size or Material",
            "example": "Size"
          },
          "isRequired": {
            "type": "boolean",
            "description": "Whether one option from this group must be selected",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "description": "Sort order",
            "example": 1
          },
          "options": {
            "example": [
              {
                "id": 200,
                "name": "80 cm",
                "priceDelta": 0,
                "sortOrder": 1
              },
              {
                "id": 201,
                "name": "120 cm",
                "priceDelta": 300,
                "sortOrder": 2
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantOptionDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "isRequired",
          "sortOrder",
          "options"
        ]
      },
...
      "BranchProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Branch product id",
            "example": 7001
          },
          "productId": {
            "type": "number",
            "description": "Product id",
            "example": 301
          },
          "categoryId": {
            "type": "number",
            "description": "Product category id",
            "example": 12
          },
          "categoryName": {
            "type": "string",
            "description": "Product category name",
            "example": "Dining Tables"
          },
          "branchId": {
            "type": "number",
            "description": "Branch id",
            "example": 5
          },
          "price": {
            "type": "number",
            "description": "Price in branch",
            "example": 25990
          },
          "stock": {
            "type": "number",
            "description": "Stock in branch",
            "example": 20
          },
          "name": {
            "type": "string",
            "description": "Product name",
            "example": "Oak Dining Table 120"
          },
          "sku": {
            "type": "string",
            "description": "Product SKU (article)",
            "example": "TBL-OAK-120"
          },
          "description": {
            "type": "string",
            "description": "Product description",
            "example": "Solid oak dining table, 120 cm width"
          },
          "slug": {
            "type": "string",
            "description": "Product slug",
            "example": "oak-dining-table-120"
          },
          "images": {
            "example": [
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/main.jpg",
                "type": "main"
              },
              {
                "url": "https://cdn.example.com/products/oak-dining-table-120/gallery-1.jpg",
                "type": "gallery"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageProductDto"
            }
          },
          "variantGroups": {
            "description": "Product variant groups. Returned in detailed endpoints by id/by-slug",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantGroupDto"
            }
          }
        },
        "required": [
          "id",
          "productId",
          "categoryId",
          "categoryName",
          "branchId",
          "price",
          "stock",
          "name",
          "slug",
          "images"
        ]
      },
      "BranchProductPaginationDto": {
        "type": "object",
        "properties": {
          "meta": {
            "description": "Pagination metadata",
            "example": {
              "total": 124,
              "page": 1,
              "limit": 25
            },
            "allOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              }
            ]
          },
          "items": {
            "example": [
              {
                "id": 7001,
                "branchId": 5,
                "price": 25990,
                "stock": 20,
                "name": "Oak Dining Table 120",
                "description": "Solid oak dining table, 120 cm width",
                "slug": "oak-dining-table-120",
                "images": [
                  {
                    "url": "https://cdn.example.com/products/oak-dining-table-120/main.jpg",
                    "type": "main"
                  }
                ]
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BranchProductDto"
            }
          }
        },
        "required": [
          "meta",
          "items"
        ]
      },
      "AdminBranchProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Branch product id",
            "example": 7001
          },
          "productId": {
            "type": "number",
            "description": "Product id",
            "example": 101
          },
          "categoryName": {
            "type": "string",
            "description": "Product category name",
            "example": "Dining Tables"
          },
          "sku": {
            "type": "string",
            "description": "Product SKU (article)",
            "example": "TBL-OAK-120"
          },
          "branchId": {
            "type": "number",
            "description": "Branch id",
            "example": 5
          },
          "price": {
            "type": "number",
            "description": "Price in branch",
            "example": 25990
          },
          "stock": {
            "type": "number",
            "description": "Stock in branch",
            "example": 20
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "description": "Created at",
            "format": "date-time",
            "example": "2026-03-10T08:45:17.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Updated at",
            "format": "date-time",
            "example": "2026-03-10T10:01:44.000Z"
          },
          "variantGroups": {
            "description": "Product variant groups. Returned in detailed admin endpoint by id",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminProductVariantGroupDto"
            }
          }
        },
        "required": [
          "id",
          "productId",
          "categoryName",
          "branchId",
          "price",
          "stock",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateBranchProductDto": {
        "type": "object",
        "properties": {
          "price": {
            "type": "number",
            "description": "Price in branch",
            "example": 24990
          },
          "stock": {
            "type": "number",
            "description": "Stock in branch",
            "example": 35
          },
          "isActive": {
            "type": "boolean",
            "description": "Active status",
            "example": true
          }
        }
      }
    }
  }
}